feat: support aria role types
