+++ packages/browser/jest-dom.d.ts
@@ -1,102 +1,6 @@
 // Disable automatic exports.
 
-
-type ARIAWidgetRole =
-    | "button"
-    | "checkbox"
-    | "gridcell"
-    | "link"
-    | "menuitem"
-    | "menuitemcheckbox"
-    | "menuitemradio"
-    | "option"
-    | "progressbar"
-    | "radio"
-    | "scrollbar"
-    | "searchbox"
-    | "slider"
-    | "spinbutton"
-    | "switch"
-    | "tab"
-    | "tabpanel"
-    | "textbox"
-    | "treeitem";
-
-type ARIACompositeWidgetRole =
-    | "combobox"
-    | "grid"
-    | "listbox"
-    | "menu"
-    | "menubar"
-    | "radiogroup"
-    | "tablist"
-    | "tree"
-    | "treegrid";
-
-type ARIADocumentStructureRole =
-    | "application"
-    | "article"
-    | "blockquote"
-    | "caption"
-    | "cell"
-    | "columnheader"
-    | "definition"
-    | "deletion"
-    | "directory"
-    | "document"
-    | "emphasis"
-    | "feed"
-    | "figure"
-    | "generic"
-    | "group"
-    | "heading"
-    | "img"
-    | "insertion"
-    | "list"
-    | "listitem"
-    | "math"
-    | "meter"
-    | "none"
-    | "note"
-    | "paragraph"
-    | "presentation"
-    | "row"
-    | "rowgroup"
-    | "rowheader"
-    | "separator"
-    | "strong"
-    | "subscript"
-    | "superscript"
-    | "table"
-    | "term"
-    | "time"
-    | "toolbar"
-    | "tooltip";
-
-type ARIALandmarkRole =
-    | "banner"
-    | "complementary"
-    | "contentinfo"
-    | "form"
-    | "main"
-    | "navigation"
-    | "region"
-    | "search";
-
-type ARIALiveRegionRole = "alert" | "log" | "marquee" | "status" | "timer";
-
-type ARIAWindowRole = "alertdialog" | "dialog";
-
-type ARIAUncategorizedRole = "code";
-
-type ARIARole =
-    | ARIAWidgetRole
-    | ARIACompositeWidgetRole
-    | ARIADocumentStructureRole
-    | ARIALandmarkRole
-    | ARIALiveRegionRole
-    | ARIAWindowRole
-    | ARIAUncategorizedRole;
+import { ARIARole } from './aria-role.ts'
 
 declare namespace matchers {
   interface TestingLibraryMatchers<E, R> {
