+++ packages/browser/rollup.config.js
@@ -141,17 +141,17 @@ export default () =>
         }),
       ],
     },
-    {
-      input: './src/client/tester/jest-dom.ts',
-      output: {
-        file: './jest-dom.d.ts',
-        format: 'esm',
-      },
-      external: [],
-      plugins: [
-        dts({
-          respectExternal: true,
-        }),
-      ],
-    },
+    // {
+    //   input: './src/client/tester/jest-dom.ts',
+    //   output: {
+    //     file: './jest-dom.d.ts',
+    //     format: 'esm',
+    //   },
+    //   external: [],
+    //   plugins: [
+    //     dts({
+    //       respectExternal: true,
+    //     }),
+    //   ],
+    // },
   ])
